Set up single test command that runs all of our tests
